Перенес отображение питомцев на страницу /pets. На страницу /main отображаю приветствующую надпись
